Link all of libc64

+++ src/static/libc64/math64.c
@@ -1,8 +1,9 @@
 #include "libc64/math64.h"
+#include "MSL_C/gcn_math.h"
 #include "MSL_C/w_math.h"
 
 f32 fatan2(f32 x, f32 y) {
-    return atan2(x, y);
+    return atan2f(x, y);
 }
 
 f32 fsqrt(f32 x) {
@@ -10,5 +11,5 @@ f32 fsqrt(f32 x) {
 }
 
 f32 facos(f32 x) {
-    return acos(x);
+    return acosf(x);
 }
